refactor: 패키지 구조 변경
Related to : #46
