+++ backend/src/main/java/com/prograngers/backend/repository/comment/CommentRepository.java
@@ -1,6 +1,6 @@
 package com.prograngers.backend.repository.comment;
 
-import com.prograngers.backend.entity.Comment;
+import com.prograngers.backend.entity.comment.Comment;
 import com.prograngers.backend.entity.solution.Solution;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
