+++ backend/src/test/java/com/prograngers/backend/fixture/CommentFixture.java
@@ -1,6 +1,6 @@
 package com.prograngers.backend.fixture;
 
-import com.prograngers.backend.entity.Comment;
+import com.prograngers.backend.entity.comment.Comment;
 import com.prograngers.backend.entity.member.Member;
 import com.prograngers.backend.entity.solution.Solution;
 import lombok.AllArgsConstructor;
